neurodebian way to install AFNI

+++ tmp/cpac_install.sh
@@ -607,6 +607,11 @@ function install_afni {
     #tar xfz ${AFNI_DOWNLOAD}.tgz
 
     wget http://fcp-indi.s3.amazonaws.com/resources/cc_afni_trusty_openmp_64.tar.gz
+    if [ $? -ne 0 ]
+    then
+        echo "unable to download AFNI from amazonws."
+        exit 1
+    fi
     tar xfz cc_afni_trusty_openmp_64.tar.gz
     rm cc_afni_trusty_openmp_64.tar.gz
     AFNI_DOWNLOAD=afni
@@ -1117,4 +1122,5 @@ do
     esac
 done
 
+echo "cd'ing back to $INIT_DIR"
 cd $INIT_DIR
